pyproject.toml
[project]
name = "my-first-blog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.2",
    "djangorestframework>=3.16.1",
    "psycopg[binary]>=3.3.3",
]
